admin

+++ src/controllers/project.controller.js
@@ -2,6 +2,17 @@
 const Model = require('../models/project.model');
 const createError = require('http-errors');
 
+// Allowed admin emails
+const ALLOWED_ADMINS = ['huy@example.com', 'huyhoang.ais@gmail.com'];
+
+// Middleware to check if user is authorized admin
+const checkAdminPermission = (req) => {
+  const userEmail = req.user?.email;
+  if (!userEmail || !ALLOWED_ADMINS.includes(userEmail)) {
+    throw createError(403, 'You do not have permission to perform this action');
+  }
+};
+
 exports.getAll = async (req, res, next) => {
   try {
     const docs = await Model.find();
@@ -19,6 +30,9 @@ exports.getById = async (req, res, next) => {
 
 exports.create = async (req, res, next) => {
   try {
+    // Check admin permission
+    checkAdminPermission(req);
+    
     const doc = await Model.create(req.body);
     res.status(201).json(doc);
   } catch (e) { next(e); }
@@ -26,6 +40,9 @@ exports.create = async (req, res, next) => {
 
 exports.updateById = async (req, res, next) => {
   try {
+    // Check admin permission
+    checkAdminPermission(req);
+    
     const doc = await Model.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true });
     if (!doc) return next(createError(404, 'Project not found'));
     res.json(doc);
@@ -34,6 +51,9 @@ exports.updateById = async (req, res, next) => {
 
 exports.removeById = async (req, res, next) => {
   try {
+    // Check admin permission
+    checkAdminPermission(req);
+    
     const doc = await Model.findByIdAndDelete(req.params.id);
     if (!doc) return next(createError(404, 'Project not found'));
     res.json({ deleted: true });
@@ -42,6 +62,9 @@ exports.removeById = async (req, res, next) => {
 
 exports.removeAll = async (req, res, next) => {
   try {
+    // Check admin permission
+    checkAdminPermission(req);
+    
     const r = await Model.deleteMany({});
     res.json({ deletedCount: r.deletedCount });
   } catch (e) { next(e); }
